refactor: move feetokens to token module
